Changes from beta 166 file

+++ Sensors/EyeOneSensor.cpp
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// Copyright (c) 2005-2008 Association Homecinema Francophone.  All rights reserved.
+// Copyright (c) 2005-2011 Association Homecinema Francophone.  All rights reserved.
 /////////////////////////////////////////////////////////////////////////////
 //
 //  This file is subject to the terms of the GNU General Public License as
@@ -621,7 +621,7 @@ CColor CEyeOneSensor::MeasureColor(COLORREF aRGBValue)
 	return EyeOneColor;
 }
 
-BOOL CEyeOneSensor::HasSpectrumCapabilities ( int * pNbBands, int * pMinWaveLength, int * pMaxWaveLength, int * pBandWidth )
+BOOL CEyeOneSensor::HasSpectrumCapabilities ( int * pNbBands, int * pMinWaveLength, int * pMaxWaveLength, double * pBandWidth )
 {
 	if ( m_CalibrationMode >= 3 )
 	{
